requirements.txt
flask==3.0.2
weaviate-client==4.5.0
openai==1.12.0
sentence-transformers==2.5.1
requests==2.31.0 